+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/controlapi/ControlApiMiddleware.java
@@ -1,21 +1,21 @@
 package com.inventage.portal.gateway.proxy.middleware.controlapi;
 
 import static com.inventage.portal.gateway.proxy.middleware.sessionBag.SessionBagMiddleware.SESSION_BAG_COOKIES;
-import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
 
 import com.inventage.portal.gateway.proxy.middleware.Middleware;
 import com.inventage.portal.gateway.proxy.middleware.oauth2.AuthenticationUserContext;
+import com.inventage.portal.gateway.proxy.middleware.sessionBag.CookieUtil;
+import io.netty.handler.codec.http.cookie.ClientCookieDecoder;
 import io.vertx.core.Handler;
 import io.vertx.core.MultiMap;
 import io.vertx.core.Vertx;
-import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.Cookie;
 import io.vertx.core.http.HttpHeaders;
 import io.vertx.ext.web.RoutingContext;
-import io.vertx.ext.web.client.HttpRequest;
 import io.vertx.ext.web.client.WebClient;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Objects;
@@ -26,31 +26,32 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Handles control api actions provided as values from a "IPS_GW_CONTROL" cookie.
+ * Handles control api actions provided as values from a "IPS_GW_CONTROL" cookie in a response's set-cookie header and also removes it.
  * Supported actions:
  * - SESSION_TERMINATE: invalidates the session and calls "end_session_endpoint" on Keycloak
- * - SESSION_RESET: resets the session by deleting all session scopes (removing all JWTs from this session)
- * and empty the session bag (removing all not whitelisted cookies).
+ * - SESSION_RESET: resets the session by deleting all oauth2 tokens (removing all JWTs from this session)
+ * and drain the session bag (removing all not keycloak related cookies).
  */
 public class ControlApiMiddleware implements Middleware {
 
     public static final String CONTROL_COOKIE_NAME = "IPS_GW_CONTROL";
-    public static final String SESSION_TERMINATE_ACTION = "SESSION_TERMINATE";
-    public static final String SESSION_RESET_ACTION = "SESSION_RESET";
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ControlApiMiddleware.class);
 
+    private static final String KEYCOAK_COOKIE_NAME_PREFIX = "KEYCLOAK_";
+    private static final String KEYCLOAK_IDENTITY_COOKIE_NAME = "KEYCLOAK_IDENTITY";
+
     private final Vertx vertx;
     private final String name;
-    private final String action;
+    private final ControlApiAction action;
     private final String resetUri;
     private final WebClient webClient;
 
-    public ControlApiMiddleware(Vertx vertx, String name, final String action, final WebClient webClient) {
+    public ControlApiMiddleware(Vertx vertx, String name, final ControlApiAction action, final WebClient webClient) {
         this(vertx, name, action, null, webClient);
     }
 
-    public ControlApiMiddleware(Vertx vertx, String name, final String action, String resetUri, final WebClient webClient) {
+    public ControlApiMiddleware(Vertx vertx, String name, final ControlApiAction action, String resetUri, final WebClient webClient) {
         this.vertx = vertx;
         this.name = name;
         this.action = action;
@@ -62,49 +63,53 @@ public class ControlApiMiddleware implements Middleware {
     public void handle(RoutingContext ctx) {
         LOGGER.debug("{}: Handling '{}'", name, ctx.request().absoluteURI());
 
-        if (ctx.session() == null) {
-            LOGGER.warn("No session initialized. Skipping session termination");
+        if ((action.equals(ControlApiAction.SESSION_TERMINATE) || action.equals(ControlApiAction.SESSION_RESET)) && ctx.session() == null) {
+            LOGGER.warn("No session initialized. Skipping session termination/reset");
             ctx.next();
             return;
         }
 
-        // on response
-        final Handler<MultiMap> respHeadersModifier = headers -> {
-            LOGGER.debug("{}: Handling response of '{}'", name, ctx.request().absoluteURI());
-
-            final Set<Cookie> storedCookies = ctx.session().get(SESSION_BAG_COOKIES);
-            if (isNotEmpty(storedCookies)) {
-                // find the first control api cookie with an action equals to the configured one
-                final List<Cookie> actionCookies = storedCookies.stream()
-                    .filter(cookie -> Objects.equals(cookie.getName(), CONTROL_COOKIE_NAME))
-                    .filter(cookie -> Objects.equals(cookie.getValue(), action))
-                    .toList();
-
-                final Optional<Cookie> controlApiActionToExecute = actionCookies.stream().findFirst();
-                if (controlApiActionToExecute.isPresent()) {
-                    LOGGER.info("Provided cookie {}", controlApiActionToExecute.get());
-                    handleAction(action, ctx);
-                }
-
-                // remove action cookies from session bag
-                actionCookies.forEach(cookieToRemove -> {
-                    if (storedCookies.remove(cookieToRemove)) {
-                        LOGGER.debug("Removing handled control api cookie from SessionBag '{}'", cookieToRemove.getName());
-                    }
-                });
-            }
-        };
+        final Handler<MultiMap> respHeadersModifier = headers -> handleControlCookies(ctx, headers);
         this.addResponseHeaderModifier(ctx, respHeadersModifier);
         ctx.next();
     }
 
-    private void handleAction(String action, RoutingContext ctx) {
+    private void handleControlCookies(RoutingContext ctx, MultiMap headers) {
+        LOGGER.debug("{}: Handling response of '{}'", name, ctx.request().absoluteURI());
+
+        final List<Cookie> cookiesToSet = headers.getAll(HttpHeaders.SET_COOKIE).stream()
+            .map(s -> ClientCookieDecoder.STRICT.decode(s))
+            .filter(cookie -> cookie != null)
+            .map(cookie -> CookieUtil.fromNettyCookie(cookie))
+            .collect(Collectors.toCollection(ArrayList::new));
+
+        final List<Cookie> actionCookies = cookiesToSet.stream()
+            .filter(cookie -> Objects.equals(cookie.getName(), CONTROL_COOKIE_NAME))
+            .filter(cookie -> Objects.equals(cookie.getValue(), action.toString()))
+            .toList();
+
+        final Optional<Cookie> controlApiActionToExecute = actionCookies.stream().findFirst();
+        if (controlApiActionToExecute.isPresent()) {
+            LOGGER.info("Provided control cookie {}", controlApiActionToExecute.get());
+            handleAction(ctx, headers, action);
+        }
+
+        final List<String> cookiesToSetWithoutActionCookies = cookiesToSet.stream()
+            .filter(cookie -> !actionCookies.contains(cookie))
+            .map(cookie -> cookie.encode())
+            .toList();
+
+        headers.remove(HttpHeaders.SET_COOKIE.toString());
+        cookiesToSetWithoutActionCookies.forEach(cookie -> headers.add(HttpHeaders.SET_COOKIE, cookie));
+    }
+
+    private void handleAction(RoutingContext ctx, MultiMap headers, ControlApiAction action) {
         switch (action) {
-            case SESSION_TERMINATE_ACTION:
+            case SESSION_TERMINATE:
                 handleSessionTermination(ctx);
                 break;
-            case SESSION_RESET_ACTION:
-                handleSessionReset(ctx);
+            case SESSION_RESET:
+                handleSessionReset(ctx, headers);
                 break;
             default:
                 final String errorMessage = "Unknown action name: '" + action + "'";
@@ -136,29 +141,15 @@ public class ControlApiMiddleware implements Middleware {
         ctx.session().destroy();
     }
 
-    private void handleSessionReset(RoutingContext ctx) {
+    private void handleSessionReset(RoutingContext ctx, MultiMap headers) {
         LOGGER.info("reset session {}", action);
 
-        informKeycloakAboutSessionReset(ctx);
-        removeAllSessionScopesFrom(ctx);
-        removeAllCookiesFromSessionBagExceptForKeycloakCookiesFrom(ctx);
+        informKeycloakAboutSessionReset(ctx, headers);
+        removeAllOauth2TokensFromSession(ctx);
+        removeAllCookiesFromSessionBagExceptForKeycloakCookies(ctx);
     }
 
-    private void removeAllSessionScopesFrom(RoutingContext ctx) {
-        AuthenticationUserContext.deleteAll(ctx.session());
-    }
-
-    private void removeAllCookiesFromSessionBagExceptForKeycloakCookiesFrom(RoutingContext ctx) {
-        LOGGER.debug("empty session bag");
-        final Set<Cookie> cookiesInSessionBag = ctx.session().get(SESSION_BAG_COOKIES);
-
-        if (cookiesInSessionBag != null) {
-            final Set<Cookie> filteredCookies = getKeycloakCookiesFrom(cookiesInSessionBag);
-            ctx.session().put(SESSION_BAG_COOKIES, filteredCookies);
-        }
-    }
-
-    private void informKeycloakAboutSessionReset(RoutingContext ctx) {
+    private void informKeycloakAboutSessionReset(RoutingContext ctx, MultiMap headers) {
         if (this.resetUri == null) {
             LOGGER.warn("Keycloak will not be informed of session reset because 'resetUri' is undefined.");
             return;
@@ -167,12 +158,10 @@ public class ControlApiMiddleware implements Middleware {
 
         try {
             final URI uri = new URI(resetUri);
-            final HttpRequest<Buffer> request = webClient.post(uri.getPort(), uri.getHost(), uri.getPath());
-            final Set<Cookie> cookies = ctx.session().get(SESSION_BAG_COOKIES);
-            final Cookie authCookie = cookies.stream().filter(cookie -> cookie.getName().equals("KEYCLOAK_IDENTITY")).findFirst().orElseThrow();
-
-            request.putHeader(HttpHeaders.COOKIE.toString(), authCookie.encode());
-            request.send()
+            webClient
+                .post(uri.getPort(), uri.getHost(), uri.getPath())
+                .putHeader(HttpHeaders.COOKIE.toString(), getKeycloakIdentityCookie(ctx, headers).encode())
+                .send()
                 .onSuccess(response -> LOGGER.info("keycloak was successfully informed of session reset. Response: {}", response))
                 .onFailure(throwable -> LOGGER.warn("keycloak was not informed of session reset: {}", throwable.getMessage()));
         } catch (URISyntaxException | NoSuchElementException e) {
@@ -180,9 +169,51 @@ public class ControlApiMiddleware implements Middleware {
         }
     }
 
+    private Cookie getKeycloakIdentityCookie(RoutingContext ctx, MultiMap headers) {
+        // 1. priority: search cookie in response header set-cookie
+        Optional<Cookie> authCookie = headers.getAll(HttpHeaders.SET_COOKIE).stream()
+            .map(s -> ClientCookieDecoder.STRICT.decode(s))
+            .filter(cookie -> cookie != null)
+            .map(cookie -> CookieUtil.fromNettyCookie(cookie))
+            .filter(cookie -> cookie.getName().equals(KEYCLOAK_IDENTITY_COOKIE_NAME))
+            .findFirst();
+
+        if (authCookie.isPresent()) {
+            return authCookie.get();
+        }
+
+        // 2. priority: search cookie in session bag
+        final Set<Cookie> cookies = ctx.session().get(SESSION_BAG_COOKIES);
+        authCookie = cookies.stream()
+            .filter(cookie -> cookie.getName().equals(KEYCLOAK_IDENTITY_COOKIE_NAME))
+            .findFirst();
+
+        if (authCookie.isPresent()) {
+            return authCookie.get();
+        }
+
+        throw new NoSuchElementException("No keycloak identity cookie");
+    }
+
+    private void removeAllOauth2TokensFromSession(RoutingContext ctx) {
+        AuthenticationUserContext.deleteAll(ctx.session());
+    }
+
+    private void removeAllCookiesFromSessionBagExceptForKeycloakCookies(RoutingContext ctx) {
+        LOGGER.debug("drain session bag");
+
+        final Set<Cookie> cookiesInSessionBag = ctx.session().get(SESSION_BAG_COOKIES);
+        if (cookiesInSessionBag == null) {
+            return;
+        }
+
+        final Set<Cookie> filteredCookies = getKeycloakCookiesFrom(cookiesInSessionBag);
+        ctx.session().put(SESSION_BAG_COOKIES, filteredCookies);
+    }
+
     private Set<Cookie> getKeycloakCookiesFrom(Set<Cookie> cookiesInSessionBag) {
         return cookiesInSessionBag.stream()
-            .filter(cookie -> cookie.getName().startsWith("KEYCLOAK_"))
+            .filter(cookie -> cookie.getName().startsWith(KEYCOAK_COOKIE_NAME_PREFIX))
             .collect(Collectors.toSet());
     }
 }