+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Publishing of the chart to the maven repository has been discontinued. Please download the chart from the helm repository instead.
 
+### Fixed
+
+- Read Control API Cookie from Response ([PORTAL-1716](https://issue.inventage.com/browse/PORTAL-1716)).
+
 ## [8.2.0]-202308021352-962-750876b6
 
 [JIRA](https://issue.inventage.com/projects/PORTAL/versions/???) - [Nexus2](https://nexus.inventage.com/#nexus-search;gav~~~8.2.0-202308021352-962-750876b6) - [Nexus3](https://nexus3.inventage.com/#browse/search=version%3D8.2.0-202308021352-962-750876b6)